Rename ux to menus and reset migrations
