+++ collectivo-test-app/test_extension/apps.py
@@ -7,7 +7,7 @@ def post_migrate_callback(sender, **kwargs):
     """Initialize extension after database is ready."""
     from collectivo.extensions.utils import register_extension
     from .populate import populate_keycloak_with_test_data
-    from collectivo.ux.utils import register_menuitem
+    from collectivo.menus.utils import register_menuitem
 
     register_extension(
         name=sender.name,
