+++ collectivo-test-app/test_extension/tests.py
@@ -2,13 +2,13 @@
 from django.test import TestCase
 from django.urls import reverse
 from rest_framework.test import APIClient
-from collectivo.ux.models import MenuItem
+from collectivo.menus.models import MenuItem
 from collectivo.extensions.models import Extension
 
 
 EXTENSIONS_URL = reverse('collectivo:collectivo.extensions:extension-list')
-MENUS_URL = reverse('collectivo:collectivo.ux:menu-list')
-ITEMS_URL = reverse('collectivo:collectivo.ux:menuitem-list',
+MENUS_URL = reverse('collectivo:collectivo.menus:menu-list')
+ITEMS_URL = reverse('collectivo:collectivo.menus:menuitem-list',
                     kwargs={'menu_id': 'main_menu'})
